requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
supabase==2.3.4
python-dotenv==1.0.0
PyYAML==6.0.1
